+++ install/migrations/update_9.5.x_to_10.0.0/schema_fixes.php
@@ -259,7 +259,7 @@ foreach ($useless_keys as $table => $keys) {
     }
 }
 
-// Add missing keys (based on glpi:tools:check_database_keys detection)
+// Add missing keys (based on tools:check_database_keys detection)
 $missing_keys = [
     'glpi_apiclients' => [
         'entities_id',