+++ src/Console/Migration/DatabasesPluginToCoreCommand.php
@@ -58,7 +58,7 @@ class DatabasesPluginToCoreCommand extends AbstractPluginToCoreCommand
     {
         parent::configure();
 
-        $this->setName('glpi:migration:databases_plugin_to_core');
+        $this->setName('migration:databases_plugin_to_core');
         $this->setDescription(__('Migrate Databases plugin data into GLPI core tables'));
     }