+++ src/Console/Migration/TimestampsCommand.php
@@ -61,7 +61,7 @@ class TimestampsCommand extends AbstractCommand
     {
         parent::configure();
 
-        $this->setName('glpi:migration:timestamps');
+        $this->setName('migration:timestamps');
         $this->setDescription(__('Convert "datetime" fields to "timestamp" to use timezones.'));
     }
 
@@ -219,7 +219,7 @@ class TimestampsCommand extends AbstractCommand
             }
             $message = sprintf(
                 __('Fix them and run the "php bin/console %1$s" command to enable timezones.'),
-                'glpi:database:enable_timezones'
+                'database:enable_timezones'
             );
             $output->writeln('<error>' . $message . '</error>', OutputInterface::VERBOSITY_QUIET);
         }